+++ src/main/java/com/codessquad/qna/web/domain/Answer.java
@@ -1,12 +1,12 @@
 package com.codessquad.qna.web.domain;
 
-import com.codessquad.qna.web.exceptions.InvalidEntityException;
+import com.codessquad.qna.web.exceptions.Entity.InvalidEntityException;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonGetter;
 
 import javax.persistence.*;
 
-import static com.codessquad.qna.web.exceptions.InvalidEntityException.EMPTY_FIELD_IN_ANSWER_ENTITY;
+import static com.codessquad.qna.web.exceptions.Entity.InvalidEntityException.EMPTY_FIELD_IN_ANSWER_ENTITY;
 import static com.codessquad.qna.web.utils.EntityCheckUtils.isNotEmpty;
 
 @Entity
