+++ src/main/java/com/codessquad/qna/web/domain/Question.java
@@ -1,6 +1,6 @@
 package com.codessquad.qna.web.domain;
 
-import com.codessquad.qna.web.exceptions.InvalidEntityException;
+import com.codessquad.qna.web.exceptions.Entity.InvalidEntityException;
 import com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import org.hibernate.annotations.Where;
@@ -8,7 +8,7 @@ import org.hibernate.annotations.Where;
 import javax.persistence.*;
 import java.util.List;
 
-import static com.codessquad.qna.web.exceptions.InvalidEntityException.EMPTY_FIELD_IN_QUESTION_ENTITY;
+import static com.codessquad.qna.web.exceptions.Entity.InvalidEntityException.EMPTY_FIELD_IN_QUESTION_ENTITY;
 import static com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException.CANNOT_MODIFY_OR_DELETE_ANOTHER_USERS_QUESTION;
 import static com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException.CAN_NOT_DELETE_BECAUSE_ANOTHER_USERS_ANSWER_IS_EXISTS;
 import static com.codessquad.qna.web.utils.EntityCheckUtils.isNotEmpty;
