+++ src/main/java/com/codessquad/qna/web/domain/User.java
@@ -1,11 +1,11 @@
 package com.codessquad.qna.web.domain;
 
-import com.codessquad.qna.web.exceptions.InvalidEntityException;
+import com.codessquad.qna.web.exceptions.Entity.InvalidEntityException;
 import com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException;
 
 import javax.persistence.*;
 
-import static com.codessquad.qna.web.exceptions.InvalidEntityException.EMPTY_FIELD_IN_USER_ENTITY;
+import static com.codessquad.qna.web.exceptions.Entity.InvalidEntityException.EMPTY_FIELD_IN_USER_ENTITY;
 import static com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException.CANNOT_MODIFY_ANOTHER_USER;
 import static com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException.PASSWORD_NOT_MATCHING;
 import static com.codessquad.qna.web.utils.EntityCheckUtils.isNotEmpty;
